backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.5.0
# PDF 파싱
pypdf>=4.3.0
pdfminer.six>=20231228
# AI / RAG
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.50.0
tiktoken>=0.7.0
# DB (Supabase REST API via httpx)
httpx>=0.27.0
# 유틸
python-multipart>=0.0.9
